test: refactor snapshot #6
