+++ packages/fresh/src/dev/builder_test.ts
@@ -5,7 +5,10 @@ import { App } from "../app.ts";
 import { DEV_ERROR_OVERLAY_URL } from "../constants.ts";
 import { BUILD_ID } from "@fresh/build-id";
 import { withTmpDir, writeFiles } from "../test_utils.ts";
-import { getStdOutput, withChildProcessServer } from "@fresh/internal/test-utils";
+import {
+  getStdOutput,
+  withChildProcessServer,
+} from "@fresh/internal/test-utils";
 import { staticFiles } from "../middlewares/static_files.ts";
 
 Deno.test({
