+++ packages/internal/README.md
@@ -1,0 +1,8 @@
+# @fresh/internal
+
+Internal-only utilities and constants for the Fresh monorepo.
+
+- Test utilities: `@fresh/internal/test-utils`
+- Version constants: `@fresh/internal/versions`
+
+Not published. Consumers outside this repo should not import these.
