+++ packages/test-utils/deno.json
@@ -1,5 +1,6 @@
 {
-  "name": "@fresh/test-utils",
+  "name": "@fresh/test-utils-DEPRECATED",
+  "private": true,
   "version": "0.1.0",
   "license": "MIT",
   "exports": {
