+++ packages/test-utils/src/mod.ts
@@ -1,7 +1,1 @@
-export * from "./process.ts";
-export * from "./browser.ts";
-export * from "./dom.ts";
-export * from "./fs.ts";
-export * from "./server.ts";
-export * from "./util.ts";
-export * from "./versions.ts";
+// Deprecated. Use @fresh/internal/test-utils instead.
